refactor: importations of meta information
 open the importation modal
  1. Drag / drop of one .csv, .tsv or .txt to NMRium drop area.
  2. import > import meta information from the left toolbar and
      then Drag / drop of one .csv, .tsv or .txt on the drop area
      in the importation modal.

close #2054
