+++ src/component/modal/metaImportation/utils/isMetaInformationFile.ts
@@ -1,0 +1,3 @@
+export function isMetaInformationFile(file: File) {
+  return ['text/csv', 'text/tsv', 'text/plain'].includes(file?.type);
+}
