+++ src/component/modal/metaImportation/utils/prepareKey.ts
@@ -1,0 +1,6 @@
+export function prepareKey(val: string | number) {
+  return String(val)
+    .toLowerCase()
+    .trim()
+    .replace(/\r?\n|\r/, '');
+}
